hms_backend/requirements.txt
Django==6.0.3
djangorestframework==3.14.0
mysqlclient==2.2.0
python-dotenv==1.0.0
google-generativeai==0.8.6
